+++ src/test/java/school/faang/user_service/service/RecommendationServiceTest.java
@@ -22,11 +22,10 @@ import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.repository.UserSkillGuaranteeRepository;
 import school.faang.user_service.repository.recommendation.RecommendationRepository;
 import school.faang.user_service.repository.recommendation.SkillOfferRepository;
+import school.faang.user_service.validator.RecommendationValidator;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -39,13 +38,15 @@ public class RecommendationServiceTest {
     @Mock
     private SkillOfferRepository skillOfferRepository;
     @Mock
-    SkillRepository skillRepository;
+    private SkillRepository skillRepository;
     @Mock
     private UserSkillGuaranteeRepository userSkillGuaranteeRepository;
     @Mock
     private UserRepository userRepository;
     @Mock
     private RecommendationMapper recommendationMapper;
+    @Mock
+    private RecommendationValidator recommendationValidator;
 
     @InjectMocks
     private RecommendationService recommendationService;
@@ -74,7 +75,7 @@ public class RecommendationServiceTest {
                 new SkillOffer(1L, new Skill(), recommendationEntity),
                 new SkillOffer(2L, new Skill(), recommendationEntity)));
 
-        when(skillRepository.existsById(anyLong())).thenReturn(true);
+        doNothing().when(recommendationValidator).validateAll(recommendationDto);
         when(recommendationMapper.toEntity(any(RecommendationDto.class))).thenReturn(recommendationEntity);
         when(recommendationMapper.toDto(any(Recommendation.class))).thenReturn(recommendationDto);
 
@@ -85,26 +86,13 @@ public class RecommendationServiceTest {
     }
 
     @Test
-    public void testCreateRecommendationWithLastUpdate() {
-        RecommendationDto recommendationDto = new RecommendationDto();
-        recommendationDto.setAuthorId(1L);
-        recommendationDto.setReceiverId(2L);
-
-        Recommendation recommendation = new Recommendation();
-        recommendation.setUpdatedAt(LocalDateTime.now().minusMonths(4));
-
-        when(recommendationRepository.findFirstByAuthorIdAndReceiverIdOrderByCreatedAtDesc(anyLong(), anyLong()))
-                .thenReturn(Optional.of(recommendation));
-
-        assertThrows(DataValidationException.class, () -> recommendationService.create(recommendationDto));
-    }
-
-    @Test
     public void testCreateRecommendationWithNullSkills() {
         RecommendationDto recommendationDto = new RecommendationDto();
         recommendationDto.setAuthorId(1L);
         recommendationDto.setReceiverId(2L);
         recommendationDto.setSkillOffers(null);
+
+        doThrow(DataValidationException.class).when(recommendationValidator).validateAll(recommendationDto);
 
         assertThrows(DataValidationException.class, () -> recommendationService.create(recommendationDto));
     }
@@ -116,22 +104,7 @@ public class RecommendationServiceTest {
         recommendationDto.setReceiverId(2L);
         recommendationDto.setSkillOffers(new ArrayList<>());
 
-        assertThrows(DataValidationException.class, () -> recommendationService.create(recommendationDto));
-    }
-
-    @Test
-    public void testCreateRecommendationWithExistIdOnSkillRepo() {
-        RecommendationDto recommendationDto = new RecommendationDto();
-        recommendationDto.setId(1L);
-        recommendationDto.setAuthorId(2L);
-        recommendationDto.setReceiverId(3L);
-        recommendationDto.setContent("content");
-        recommendationDto.setSkillOffers(List.of(
-                new SkillOfferDto(1L, 1L),
-                new SkillOfferDto(2L, 2L)));
-
-        when(skillRepository.existsById(1L)).thenReturn(true);
-        when(skillRepository.existsById(2L)).thenReturn(false);
+        doThrow(DataValidationException.class).when(recommendationValidator).validateAll(recommendationDto);
 
         assertThrows(DataValidationException.class, () -> recommendationService.create(recommendationDto));
     }
@@ -139,9 +112,8 @@ public class RecommendationServiceTest {
     @Test
     public void testUpdateWithValidateException() {
         RecommendationDto recommendationDto = new RecommendationDto();
-        recommendationDto.setId(1L);
 
-        when(recommendationRepository.findById(1L)).thenReturn(Optional.empty());
+        doThrow(DataValidationException.class).when(recommendationValidator).validateRecommendationForUpdate(recommendationDto);
 
         assertThrows(DataValidationException.class, () -> recommendationService.update(recommendationDto));
     }
@@ -170,8 +142,8 @@ public class RecommendationServiceTest {
                 new SkillOffer(1L, new Skill(), recommendationEntity),
                 new SkillOffer(2L, new Skill(), recommendationEntity)));
 
-        when(recommendationRepository.findById(1L)).thenReturn(Optional.of(recommendationEntity));
-        when(skillRepository.existsById(anyLong())).thenReturn(true);
+        doNothing().when(recommendationValidator).validateRecommendationForUpdate(recommendationDto);
+        doNothing().when(recommendationValidator).validateAll(recommendationDto);
         when(recommendationMapper.toEntity(any(RecommendationDto.class))).thenReturn(recommendationEntity);
         when(recommendationMapper.toDto(any(Recommendation.class))).thenReturn(recommendationDto);
 
@@ -222,7 +194,7 @@ public class RecommendationServiceTest {
     }
 
     @Test
-    public void testAllRecommendation() {
+    public void testGetAllRecommendation() {
         long receiverId = 1L;
         int pageNumber = 0;
         int pageSize = 10;
